+++ examples/hello/clients/hello-nested/hello-nested.nix
@@ -7,7 +7,7 @@ base.languages.python.mkClient {
   version = "1.0.0";
   src = ./.;
   pythonVersion = python3;
-  # Here we don't use pp with numpyWrapper since it's our own
+  # Here we just use numpyWrapper since it's our own
   # package and not part of the python version packages.
-  propagatedBuildInputs = (pp: [ numpyWrapper.package ]);
+  propagatedBuildInputs = (_: [ numpyWrapper.package ]);
 }