+++ languages/default.nix
@@ -2,7 +2,7 @@ pkgs: base: versions:
 let
   all = rec {
     rust = pkgs.callPackage ./rust { inherit base versions; };
-    python = pkgs.callPackage ./python { inherit base versions; };
+    python = pkgs.callPackage ./python { inherit base; };
     terraform = pkgs.callPackage ./terraform { inherit base versions; };
     markdown = pkgs.callPackage ./markdown { inherit base; };
   };