+++ overlays/wasilibc.nix
@@ -1,9 +1,9 @@
-versions: self: super:
 let
   # set this to true to produce a wasilibc with debug symbols
   # TODO: there is probably a better place for this
   debug = false;
 in
+versions: self: super:
 {
   # create statically built versions of the llvm libraries (wasi does not support dynamic linking)
   llvmPackages_12 = super.llvmPackages_12 // {