+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   explicitly define names of unfree packages that can be installed despite being unfree
   for a project when importing Nedryland. 
 - Shellcheck is now available as part of the `ci` component.
+- nix-lint is now available as part of the `ci` component.
 
 ## [1.2.0] - 2021-08-31